refactor: 재고 관련 로직 상품 로직 병합
