+++ apps/commerce-api/src/main/java/com/loopers/domain/product/ProductRepository.java
@@ -4,10 +4,16 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
 public interface ProductRepository {
 
     Product findById(Long id);
 
     Page<Product> search(Long brandId, Pageable pageable, ProductSortType sortType);
+
+    Stock findStockByProductId(Long productId);
+
+    List<Stock> findStocksByProductIds(List<Long> productIds);
 }
