--- apps/commerce-api/src/main/java/com/loopers/domain/product/StockRepository.java
@@ -1,13 +1,0 @@
-package com.loopers.domain.product;
-
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-
-@Repository
-public interface StockRepository {
-
-    Stock findByProductId(Long productId);
-
-    List<Stock> findAllByProductIds(List<Long> productIds);
-}
