--- apps/commerce-api/src/main/java/com/loopers/domain/product/StockService.java
@@ -1,21 +1,0 @@
-package com.loopers.domain.product;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-
-@Service
-@RequiredArgsConstructor
-public class StockService {
-
-    private final StockRepository stockRepository;
-
-    public Stock getDetailByProductId(Long productId) {
-        return stockRepository.findByProductId(productId);
-    }
-
-    public List<Stock> getListByProductIds(List<Long> productIds) {
-        return stockRepository.findAllByProductIds(productIds);
-    }
-}
